requirements.txt
torch==1.11.0
torchvision==0.12.0
torchaudio==0.11.0
opencv-python==4.6.0.66
lpips==0.1.4
face_alignment==1.3.5
kornia==0.6.7
matplotlib
flask
gunicorn
google-cloud-storage